requirements.txt
langchain
beautifulsoup4
python-dotenv
tinydb